refactor: centralize ruleset conversion
